public/images/3. gpt_integrate_frontend/3. gpt_integrate_frontend/fastgpt/pyproject.toml
[tool.poetry]
name = "fastgpt"
version = "0.1.0"
description = ""
authors = ["asif-langrah <ak.development.pk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
openai = "^1.33.0"
python-dotenv = "^1.0.1"
sqlmodel = "^0.0.19"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
